refactor(core): colocate claim aggregate builder
